Initial commit
I am gonna regret not refactoring this who-knows-how-many years old
project before continuing to work on it, but hopefully things are not as
bad as they seem. I will add entries to .gitignore, update libraries,
refactor the code, add gradle tasks for texture packing, restructure the
whole thing so that assets is not in android, and all of that
later, right now I want to at least push this as-is.